Change xml unmarshal strategy for non-trivial builtins
Work-in-progress. For non-trivial builtins like dateTime, gDay,
base64Binary, etc, previously we would generate a non-exported type
declaration with the corresponding `UnmarshalText`/`MarshalText`
declarations. For example,

	type xsdDate time.Time

	func (t *xsdDate) UnmarshalText(text []byte) error {
		return _unmarshalTime(text, (*time.Time)(t), "2006-01-02")
	}

	type Library struct {
		Book      []Book  `xml:"http://www.example.com/ book"`
		Title     string  `xml:"http://www.example.com/ title"`
		Published xsdDate `xml:"http://www.example.com/ published"`
		Author    string  `xml:"http://www.example.com/ author"`
	}

The problem with this approach is that it exposes a private type in
an exported struct. It's not clear that the `Published` field above can
be converted to a `time.Time` type, and it's not clear how to set that
field. This detail should be hidden from the user. This is the ideal type
declaration:

	type Library struct {
		Book      []Book  `xml:"http://www.example.com/ book"`
		Title     string  `xml:"http://www.example.com/ title"`
		Published time.Time `xml:"http://www.example.com/ published"`
		Author    string  `xml:"http://www.example.com/ author"`
	}

This commit does just that by adding MarshalXML and UnmarshalXML methods
that "do the right thing". It also reworks how existing Mashal/Unmarshal text
methods work; unexported helper types are added for each non-trivial builtin,
that are leveraged by the marshal/unmarshal methods for both simple and
complex types.