Add marshal/unmarshal tests for generated Go code.
To add a new test, create two files in xsdgen/testdata:
- <name>.xsd , containing a single XML schema document.
- <name>.xml, containing sample data conforming to the schema.

Then run 'go generate' in the xsdgen directory. The file name should
be a valid Go identifier, and the schema should list as a top-level element
every element in the sample data.

The file <name>_gen_test.go will be generated, containing type declarations
from the schema, and a unit test function, Test<Name>, that does the following:

- Declares a top-level Document struct containing each element described
  in the schema.
- Unmarshals the sample file into said struct.
- Marshals said struct back into XML.
- Confirms that the output XML is equivalent to the input XML.

The `xmltree.Equal` method was added for this purpose; it disregards order,
leading/trailing white space, and different namespace prefixes when comparing
two documents.