Refactor/revamp handling of references
The main change with this commit is the use of the `internal/dependency`
package to resolve references such as those in

		<complexType name="foo">
		  <attribute ref="xml:lang" />
		  <attrGroup ref="commonAttributes" />
		</complexType>

in a more "correct" way. Previously, the `xsd` package could have done
the wrong thing when flattening long chains of references. Note that
flattening references is not 100% correct, since the xsd spec technically
allows for certain kinds of reference loops, but flattening all references
greatly simplifies parsing.

This commit also fixes a bug when generating names for anonymous types,
that caused them to be named twice and left unknown types in the
`memberType` attributes of unions.