Name elements before attributes in structs
Commit d363344 added code to avoid name collisions between
struct fields for type elements vs attributes. It had code to use the
suffix "Attr" as a possible differentiator for attributes, before falling
through to using numbers. But this code path was never exercised
because attributes were always added to the struct first.

This change names elements first, then attributes second. Granting
the original name to elements over attributes is an arbitrary decision
on my part, because I feel that struct fields are already more similar
to sub-elements within an XML object when compared to attributes.